backend/requirements.txt
Django
djangorestframework
drf_yasg
pytest
ruff